fix(sdk): address comments including move of API helpers to KubernetesBackend
